Split form controller into public and internal parts
