+++ lib/keila_web/views/public_form_layout_view.ex
@@ -1,4 +1,4 @@
-defmodule KeilaWeb.FormLayoutView do
+defmodule KeilaWeb.PublicFormLayoutView do
   use KeilaWeb, :view
   use Phoenix.HTML
